Added user login, userId in Ads

+++ API/Controllers/V1/Ads/AdsController.cs
@@ -1,16 +1,21 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Threading.Tasks;
 using API.Contracts.V1;
 using API.Domain;
+using API.Extensions;
 using API.Models.Ad;
 using API.Services;
 using AutoMapper;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers.V1.Ads
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class AdsController : ControllerBase
     {
 
@@ -26,21 +31,23 @@ namespace API.Controllers.V1.Ads
         [HttpPost(ApiRoutes.Ads.Create)]
         public async Task<IActionResult> Create([FromForm] AdCreationModel model)
         {
+            
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
             
             var create = _mapper.Map<Ad>(model);
-
-            await _adService.CreateAdAsync(create);
-
+        
+            await _adService.CreateAdAsync(HttpContext.GetUserId(), create);
+        
             var result = _mapper.Map<AdDetailsModel>(create);
-
-            return CreatedAtAction("Get", new {adId = create.Id}, result);
+        
+            return CreatedAtAction("Get", new {adId = create.Id, userId = create.UserId}, result);
         }
 
         [HttpGet(ApiRoutes.Ads.Get)]
+        [AllowAnonymous]
         public async Task<IActionResult> Get([FromRoute] Guid adId)
         {
                 var ad = await _adService.GetAdByIdAsync(adId);
@@ -53,6 +60,7 @@ namespace API.Controllers.V1.Ads
         }
 
         [HttpGet(ApiRoutes.Ads.GetAll)]
+        [AllowAnonymous]
         public async Task<IActionResult> GetAll()
         {
             var ads = await _adService.GetAdsAsync();
@@ -66,6 +74,7 @@ namespace API.Controllers.V1.Ads
         }
 
         [HttpGet(ApiRoutes.Ads.GetByCategory)]
+        [AllowAnonymous]
         public async Task<IActionResult> GetByCategory([FromRoute]Guid categoryId)
         {
             var ads = await _adService.GetAdsByCategory(categoryId);
@@ -81,8 +90,16 @@ namespace API.Controllers.V1.Ads
         [HttpPut(ApiRoutes.Ads.Update)]
         public async Task<IActionResult> Update([FromForm] AdUpdateModel model)
         {
+            
+            var userOwnsPost = await _adService.UserOwnsPostAsync(model.Id, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post" });
+            }
 
-            var result = await _adService.UpdateAdAsync(model);
+            var ad = await _adService.GetAdByIdAsync(model.Id);
+            
+            var result = await _adService.UpdateAdAsync(_mapper.Map<AdUpdateModel>(ad));
             
             if (result)
             {
@@ -96,6 +113,13 @@ namespace API.Controllers.V1.Ads
         [HttpDelete(ApiRoutes.Ads.Delete)]
         public async Task<IActionResult> Delete([FromRoute]Guid adId)
         {
+            var userOwnsPost = await _adService.UserOwnsPostAsync(adId, HttpContext.GetUserId());
+            
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post" });
+            }
+            
             var deleted = await _adService.DeleteAdAsync(adId);
 
             if (deleted)
@@ -105,6 +129,8 @@ namespace API.Controllers.V1.Ads
 
             return NotFound("Such ad does not exists.");
         }
+        
+        
         
     }
 }
